Introduce logging mechanism & some initial logs
The logging is based on the LoggerCategory nested classes to define the
log category (currently only Parser) and on defining LoggerMessage in
the LoggerParserExtensions class (mainly the enum ParserEventId for the
Parser category log).
In some case, creating a scope will also log a message to indicating
entering in a "critical" zone.